refactor(application-config): rename validations-config file, use named exports
